As ShelleyBasedEra is actually a feature, move it to the Feature parent module
