+++ cardano-api/internal/Cardano/Api/Block.hs
@@ -51,7 +51,9 @@ module Cardano.Api.Block (
     makeChainTip,
   ) where
 
+import           Cardano.Api.Eon.ShelleyBasedEra
 import           Cardano.Api.Eras
+import           Cardano.Api.Eras.Constraints
 import           Cardano.Api.Hash
 import           Cardano.Api.HasTypeProxy
 import           Cardano.Api.Keys.Shelley
