+++ cardano-api/internal/Cardano/Api/Convenience/Construction.hs
@@ -17,7 +17,7 @@ module Cardano.Api.Convenience.Construction (
 
 import           Cardano.Api.Address
 import           Cardano.Api.Certificate
-import           Cardano.Api.Eras
+import           Cardano.Api.Eon.ShelleyBasedEra
 import           Cardano.Api.Fees
 import           Cardano.Api.ProtocolParameters
 import           Cardano.Api.Query
