+++ cardano-api/internal/Cardano/Api/LedgerState.hs
@@ -83,7 +83,9 @@ module Cardano.Api.LedgerState
 
 import           Cardano.Api.Block
 import           Cardano.Api.Certificate
-import           Cardano.Api.Eras
+import           Cardano.Api.Eon.ShelleyBasedEra
+import           Cardano.Api.Eras.Case
+import           Cardano.Api.Eras.Constraints
 import           Cardano.Api.Error
 import           Cardano.Api.Genesis
 import           Cardano.Api.IO
