+++ cardano-api/internal/Cardano/Api/Query/Types.hs
@@ -14,7 +14,7 @@ module Cardano.Api.Query.Types
   , toDebugLedgerStatePair
   ) where
 
-import           Cardano.Api.Eras.Core
+import           Cardano.Api.Eon.ShelleyBasedEra
 
 import           Cardano.Binary
 import           Cardano.Ledger.Binary
