+++ cardano-api/src/Cardano/Api/Shelley.hs
@@ -298,7 +298,7 @@ import           Cardano.Api.Address
 import           Cardano.Api.Block
 import           Cardano.Api.Certificate
 import           Cardano.Api.DRepMetadata
-import           Cardano.Api.Eras
+import           Cardano.Api.Eon.ShelleyBasedEra
 import           Cardano.Api.Genesis
 import           Cardano.Api.Governance.Actions.ProposalProcedure
 import           Cardano.Api.Governance.Actions.VotingProcedure
